+++ packages/yourloops/lib/patient/models/patient-context-result.model.ts
@@ -25,22 +25,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { type PatientFilterStats } from './patient-filter-stats.model'
 import { type Team } from '../../team'
 import { type MedicalData } from '../../data/models/medical-data.model'
 import { type Patient } from './patient.model'
-import { type PatientListFilters } from '../../../components/patient-list/enums/patient-list.enum'
 import { type PatientTeam } from './patient-team.model'
 
 export interface PatientContextResult {
   patients: Patient[]
-  patientsFilterStats: PatientFilterStats
-  pendingPatientsCount: number | undefined
+  pendingPatientsCount?: number
+  allPatientsForSelectedTeamCount?: number
   initialized: boolean
   refreshInProgress: boolean
   getPatientByEmail: (email: string) => Patient
   getPatientById: (userId: string) => Patient
-  filterPatients: (filterType: PatientListFilters, search: string, flaggedPatients: string[]) => Patient[]
+  searchPatients: (search: string) => Patient[]
   invitePatient: (team: Team, username: string) => Promise<void>
   editPatientRemoteMonitoring: (patient: Patient) => void
   markPatientMessagesAsRead: (patient: Patient) => void