+++ packages/yourloops/lib/patient/patient.util.ts
@@ -31,8 +31,8 @@ import moment from 'moment-timezone'
 import { type Patient } from './models/patient.model'
 import { type PatientTeam } from './models/patient-team.model'
 import { UserInvitationStatus } from '../team/models/enums/user-invitation-status.enum'
-import { PatientListFilters } from '../../components/patient-list/enums/patient-list.enum'
 import { type User } from '../auth'
+import { type PatientsFilters } from '../filter/models/patients-filters.model'
 
 export default class PatientUtils {
   static removeDuplicates(patientsWithDuplicates: Patient[]): Patient[] {
@@ -94,6 +94,10 @@ export default class PatientUtils {
     })
   }
 
+  static isInvitationPending = (patient: Patient, selectedTeamId: string): boolean => {
+    return patient.teams.some((team: PatientTeam) => team.teamId === selectedTeamId && team.status === UserInvitationStatus.pending)
+  }
+
   static getNonPendingPatients = (patients: Patient[], selectedTeamId: string): Patient[] => {
     return patients.filter(patient => patient.teams.some(team => team.teamId === selectedTeamId && team.status !== UserInvitationStatus.pending))
   }
@@ -102,38 +106,42 @@ export default class PatientUtils {
     return patients.filter(patient => patient.teams.some(team => team.teamId === selectedTeamId && team.status === UserInvitationStatus.pending))
   }
 
-  static isInTeam = (patient: Patient, teamId: string): boolean => {
-    const tm = patient.teams.find((team: PatientTeam) => team.teamId === teamId)
-    return !!tm
+  static filterPatientsOnMonitoringAlerts = (patients: Patient[], patientFilters: PatientsFilters): Patient[] => {
+    if (!patientFilters.timeOutOfTargetEnabled && !patientFilters.hypoglycemiaEnabled && !patientFilters.dataNotTransferredEnabled) {
+      return patients
+    }
+    if (patientFilters.timeOutOfTargetEnabled && patientFilters.hypoglycemiaEnabled && patientFilters.dataNotTransferredEnabled) {
+      return patients.filter(patient => patient.alarms?.timeSpentAwayFromTargetActive || patient.alarms?.frequencyOfSevereHypoglycemiaActive || patient.alarms?.nonDataTransmissionActive)
+    }
+    if (patientFilters.timeOutOfTargetEnabled && patientFilters.hypoglycemiaEnabled) {
+      return patients.filter(patient => patient.alarms?.timeSpentAwayFromTargetActive || patient.alarms?.frequencyOfSevereHypoglycemiaActive)
+    }
+    if (patientFilters.hypoglycemiaEnabled && patientFilters.dataNotTransferredEnabled) {
+      return patients.filter(patient => patient.alarms?.frequencyOfSevereHypoglycemiaActive || patient.alarms?.nonDataTransmissionActive)
+    }
+    if (patientFilters.timeOutOfTargetEnabled && patientFilters.dataNotTransferredEnabled) {
+      return patients.filter(patient => patient.alarms?.timeSpentAwayFromTargetActive || patient.alarms?.nonDataTransmissionActive)
+    }
+    if (patientFilters.timeOutOfTargetEnabled) {
+      return patients.filter(patient => patient.alarms?.timeSpentAwayFromTargetActive)
+    }
+    if (patientFilters.hypoglycemiaEnabled) {
+      return patients.filter(patient => patient.alarms?.frequencyOfSevereHypoglycemiaActive)
+    }
+    return patients.filter(patient => patient.alarms?.nonDataTransmissionActive)
   }
 
-  static extractPatients = (patients: Patient[], filterType: PatientListFilters, flaggedPatients: string[], selectedTeamId?: string): Patient[] => {
-    const twoWeeksFromNow = new Date()
-    switch (filterType) {
-      case PatientListFilters.All:
-        return selectedTeamId ? PatientUtils.getNonPendingPatients(patients, selectedTeamId) : patients
-      case PatientListFilters.Pending:
-        return PatientUtils.getPendingPatients(patients, selectedTeamId)
-      case PatientListFilters.Flagged:
-        return patients.filter(patient => flaggedPatients.includes(patient.userid))
-      case PatientListFilters.UnreadMessages:
-        return patients.filter(patient => patient.metadata.hasSentUnreadMessages)
-      case PatientListFilters.OutOfRange:
-        return patients.filter(patient => patient.alarms.timeSpentAwayFromTargetActive)
-      case PatientListFilters.SevereHypoglycemia:
-        return patients.filter(patient => patient.alarms.frequencyOfSevereHypoglycemiaActive)
-      case PatientListFilters.DataNotTransferred:
-        return patients.filter(patient => patient.alarms.nonDataTransmissionActive)
-      case PatientListFilters.RemoteMonitored:
-        return patients.filter(patient => patient.monitoring?.enabled)
-      case PatientListFilters.Private:
-        return patients.filter(patient => PatientUtils.isInTeam(patient, filterType))
-      case PatientListFilters.Renew:
-        twoWeeksFromNow.setDate(twoWeeksFromNow.getDate() + 14)
-        return patients.filter(patient => patient.monitoring?.enabled && patient.monitoring.monitoringEnd && new Date(patient.monitoring.monitoringEnd).getTime() - twoWeeksFromNow.getTime() < 0)
-      default:
-        return patients
+  static extractPatients = (patients: Patient[], patientFilters: PatientsFilters, flaggedPatientsId: string[] | undefined, selectedTeamId: string): Patient[] => {
+    // When the filter is pending, we only get the pending patients and don't apply any filter on them
+    if (patientFilters.pendingEnabled) {
+      return patients.filter((patient) => PatientUtils.isInvitationPending(patient, selectedTeamId))
     }
+    // We do not take the pending patients
+    const nonPendingPatients = PatientUtils.getNonPendingPatients(patients, selectedTeamId)
+    return PatientUtils.filterPatientsOnMonitoringAlerts(nonPendingPatients, patientFilters)
+      .filter(patient => patientFilters.telemonitoredEnabled ? patient.monitoring?.enabled : patient)
+      .filter(patient => patientFilters.manualFlagEnabled ? flaggedPatientsId?.includes(patient.userid) : patient)
+      .filter(patient => patientFilters.messagesEnabled ? patient.metadata.hasSentUnreadMessages : patient)
   }
 
   static extractPatientsWithBirthdate = (patients: Patient[], birthdate: string, firstNameOrLastName: string): Patient[] => {