+++ packages/yourloops/test/integration/pages/notification/notification-for-patient.spec.tsx
@@ -28,9 +28,9 @@
 import { waitFor } from '@testing-library/react'
 import { renderPage } from '../../utils/render'
 import { mockPatientLogin } from '../../mock/patient-login.mock'
-import { monitoredPatientAsTeamMember } from '../../mock/patient.api.mock'
+import { monitoredPatientAsTeamMember } from '../../data/patient.api.data'
 import NotificationApi from '../../../../lib/notifications/notification.api'
-import { iTeamOne, teamOne, teamTwo } from '../../mock/team.api.mock'
+import { iTeamOne, buildTeamOne, buildTeamTwo } from '../../mock/team.api.mock'
 import TeamAPI from '../../../../lib/team/team.api'
 import { invitationTeam } from '../../data/notification.data'
 import { checkAcceptTeamInvite, closeDialogNotificationTeam } from '../../assert/notification-join-team'
@@ -38,7 +38,7 @@ import { checkAcceptTeamInvite, closeDialogNotificationTeam } from '../../assert
 describe('Notification page for patient', () => {
   beforeAll(() => {
     mockPatientLogin(monitoredPatientAsTeamMember)
-    jest.spyOn(TeamAPI, 'getTeams').mockResolvedValue([teamOne, teamTwo])
+    jest.spyOn(TeamAPI, 'getTeams').mockResolvedValue([buildTeamOne(), buildTeamTwo()])
     jest.spyOn(NotificationApi, 'getReceivedInvitations').mockResolvedValue([invitationTeam])
     jest.spyOn(TeamAPI, 'getTeamFromCode').mockResolvedValue(iTeamOne)
     jest.spyOn(NotificationApi, 'acceptInvitation').mockResolvedValue()